Fix issues with users on MacOS.
MacOS is weird. the bind directory is using a driver where the directory is mode 755, the owner and
group for the directory are both root, yet the app id 1000 can write to the directory.

Explicitly check for that condition, and assume that if we can write a file as the app user that
the permissions are already magically correct and leave all the ids alone.